requirements.txt
djangorestframework>=3.7.7
Django>=2.0.1
pygments
coreapi
django-cors-headers
django-filter